requirements.txt
lyricsgenius
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy
nltk
gensim
python-dotenv